test: data box tests
